pyproject.toml
[project]
name = "virtus-bot"
version = "0.1.0"
description = "A Discord bot for managing user experience points"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.3",
    "fastapi>=0.128.0",
    "uvicorn>=0.39.0",
    "aiohttp>=3.12.12",
]
